requirements.txt
openai>=1.0.0
sentence-transformers>=2.2.2
numpy>=1.24.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
tqdm>=4.65.0
